requirements.txt
boto3==1.26.113
botocore==1.29.113
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
python-whois==0.8.0
cymruwhois==1.6
tqdm==4.65.0
pyfiglet==0.8.post1
requests==2.31.0
click==8.1.3
tabulate==0.9.0
shodan==1.30.0
google_cloud_logging-3.7.0
